change router structure
-separate router
-remove unused js files
