--- routes/booking_app_routes/category_selector.js
@@ -1,36 +1,0 @@
-var express = require('express');
-var router = express.Router();
-
-
-/* GET home page. */
-router.get('/', (req, res, next) => {
-    // MongoDB에 요청해 json 형태로 데이터를 가져와
-    // 그 데이터를 바로 전송
-  
-    // categoryDatas = ;
-    // var data = []
-    //   categoryDatas.forEach(function(object){
-    //     data.push({
-    //       name : 'Dummy',
-    //       color: '#1abc9c'
-    //     });
-    //   });
-  
-    // 클라이언트에 데이터를 json으로 쏨
-    res.json({
-            name : 'Dummy',
-            color: '#1abc9c'
-          })
-
-
-});
-
-router.get('/second', (req, res, next) => {
-    res.json({
-      name : 'Second',
-      color: '#1abc9c'
-    })
-});
-
-
-module.exports = router;
