+++ routes/booking_web_routes/index.js
@@ -1,0 +1,10 @@
+var express = require('express');
+var router = express.Router();
+
+router.post("/", (req, res, next) => {
+
+})
+
+router.get("/", (req, res, next) => {
+
+})
